Server-Side Event Listerners and Chat Components Set Up
The applications socket.io library was set up in the backend and tested on the frontend. The data communication was flawless. The various components that make up the chat user-interface was then set up and some further developed. On the server-side, I send up helper functions and also the routes to for the api. No database has been set up yet.